rename fusing mateadata wrapper
