+++ model_compression_toolkit/core/pytorch/back2framework/pytorch_model_builder.py
@@ -23,7 +23,7 @@ from networkx import topological_sort
 
 from model_compression_toolkit.core import FrameworkInfo
 from model_compression_toolkit.core import common
-from model_compression_toolkit.core.common.fusion.graph_with_fusing_metadata import GraphWithFusingMetadata
+from model_compression_toolkit.core.common.fusion.fusing_metadata_wrapper import FusingMetadataWrapper
 from model_compression_toolkit.logger import Logger
 from model_compression_toolkit.core.common import BaseNode, Graph
 from model_compression_toolkit.core.common.back2framework.base_model_builder import BaseModelBuilder
@@ -229,10 +229,10 @@ class PytorchModel(torch.nn.Module):
         self.graph = copy.deepcopy(graph)
         if isinstance(graph, Graph): # Pruning still uses Graph and not GraphWithFusingMetadata
             delattr(self.graph, 'fqc')
-        elif isinstance(graph, GraphWithFusingMetadata):
+        elif isinstance(graph, FusingMetadataWrapper):
             delattr(self.graph.get_internal_graph(), 'fqc')
         else:
-            raise ValueError(f"The graph should be of type {type(Graph)} or {type(GraphWithFusingMetadata)} but is {type(graph)}.")
+            raise ValueError(f"The graph should be of type {type(Graph)} or {type(FusingMetadataWrapper)} but is {type(graph)}.")
 
         self.node_sort = list(topological_sort(self.graph))
         self.node_to_activation_quantization_holder = {}
