+++ tests_pytest/_fw_tests_common_base/base_fusion_metadata_wrapper_test.py
