+++ tests_pytest/pytorch_tests/integration_tests/core/test_fusion_metadata_wrapper_torch.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 
 
-from tests_pytest._fw_tests_common_base.base_graph_with_fusion_metadata_test import BaseGraphWithFusingMetadataTest
+from tests_pytest._fw_tests_common_base.base_fusion_metadata_wrapper_test import BaseGraphWithFusingMetadataTest
 from tests_pytest.pytorch_tests.torch_test_util.torch_test_mixin import TorchFwMixin
 
 import torch.nn as nn
